Add links to all the things

+++ govtrack/views.py
@@ -1,7 +1,8 @@
-from django.shortcuts import get_object_or_404, render, redirect, Http404
+from django.shortcuts import get_object_or_404, render, redirect, Http404, HttpResponse
+from django.forms import formset_factory
 
-from .models import Declaration, Country, Node, NodeType
-from .forms import NodeTypeForm, NodeForm, DeclarationForm
+from .models import Declaration, Country, Node, NodeType, Link
+from .forms import NodeTypeForm, NodeForm, DeclarationForm, LinkForm, CountryForm
 
 import logging 
 
@@ -31,6 +32,7 @@ def node(request, node_id):
         'record': node,
         'country': node.country,
         'parents_list': node.ancestors,
+        'links': node.links.all(),
     })
 
 def countries(request):
@@ -39,8 +41,30 @@ def countries(request):
         c.node_population = c.get_root_node().declared_population()
     return render(request, 'govtrack/countries.html', {'country_list': clist})
 
-def country(request, country_id):
+def country(request, country_id, action='view'):
     country = get_object_or_404(Country, pk=country_id)
+    form = CountryForm(instance=country)
+
+    link_initial = {
+        'content_type': Country.content_type_id(),
+        'object_id': country_id,
+    }
+    linkform = LinkForm(initial=link_initial)
+    
+    # If POST received, save form
+    if request.method == 'POST':
+        form = CountryForm(request.POST, instance=country)
+        linkform = LinkForm(request.POST, initial=link_initial)
+        do_redir = False
+        if form.is_valid():
+            saved = form.save()
+            action='view'
+            if linkform.has_changed():
+                if linkform.is_valid():
+                    linkform.save()
+                else:
+                    print("did not save url because %s " % linkform.errors)
+                    action='edit'
 
     structure = country.get_root_nodetype().build_hierarchy()
     records = country.get_root_node().build_hierarchy()
@@ -53,22 +77,43 @@ def country(request, country_id):
         
 
     return render(request, 'govtrack/country.html', {
+        'action': action,
         'country': country,
         'structure_list': structure,
         'records_list': records,
-        'total_declared_population': country.get_root_node().declared_population()
+        'total_declared_population': country.get_root_node().declared_population(),
+        'links': country.links.all(),
+        'form': form,
+        'linkform': linkform,
         })
 
 
 def nodetype_edit(request, nodetype_id):
     nodetype = get_object_or_404(NodeType, pk=nodetype_id)
     form = NodeTypeForm(instance=nodetype)
+    link_initial = {
+        'content_type': NodeType.content_type_id(),
+        'object_id': nodetype_id,
+    }
     if request.method == 'POST':
         form = NodeTypeForm(request.POST, instance=nodetype)
+        linkform = LinkForm(request.POST, initial=link_initial)
         if form.is_valid():
             form.save()
+            if linkform.has_changed() and linkform.is_valid():
+                linkform.save()
         return redirect('country', country_id=nodetype.country.id)
-    return render(request, 'govtrack/nodetype.html', {'action': 'edit', 'nodetype': nodetype, 'form': form, 'country': nodetype.country, 'parents_list': nodetype.ancestors})
+
+    linkform = LinkForm(initial=link_initial)
+    return render(request, 'govtrack/nodetype.html', {
+        'action': 'edit',
+        'nodetype': nodetype,
+        'form': form,
+        'country': nodetype.country,
+        'parents_list': nodetype.ancestors,
+        'links': nodetype.links.all(),
+        'linkform': linkform,
+        })
 
 def nodetype_child(request, parent_id):
     parent = get_object_or_404(NodeType, pk=parent_id)
@@ -88,26 +133,43 @@ def nodetype_child(request, parent_id):
     return render(request, 'govtrack/nodetype.html', {'action': 'add', 'form': form, 'country': parent.country, 'parent': parent, 'parents_list': parent.ancestors})
 
 def node_edit(request, node_id):
-    node = Node.objects.get(id=node_id)
-    if not node:
-        raise Http404("No such node")
+    node = get_object_or_404(Node, pk=node_id)
     form = NodeForm(instance=node)
     form.fields['supplements'].queryset = node.parent.get_supplement_choices()
 
+    link_initial = {
+        'content_type': Node.content_type_id(),
+        'object_id': node_id,
+    }
+    linkform = LinkForm(initial=link_initial)
     # If POST received, save form
     if request.method == 'POST':
         form = NodeForm(request.POST, instance=node)
+        linkform = LinkForm(request.POST, initial=link_initial)
+        do_redir = False
         if form.is_valid():
+            do_redir=True
             saved = form.save()
-            return redirect('country', country_id=node.country.id)
+            if linkform.has_changed():
+                do_redir=False
+                if linkform.is_valid():
+                    linkform.save()
+                    do_redir=True
+                else:
+                    print("did not save url because %s " % linkform.errors)
+            if do_redir:
+                return redirect('country', country_id=node.country.id)
     # Show form
-    nodedata = {
-        'name': '',
-        'parent': node.parent.id,
-        'country': node.parent.country.id,
-        'nodetype': node.nodetype.id
-    }
-    return render(request, 'govtrack/node.html', {'action': 'edit', 'record': node, 'form': form, 'country': node.country, 'parents_list': node.ancestors, 'supplements_list': node.supplements.all()})
+    return render(request, 'govtrack/node.html', {
+        'action': 'edit',
+        'record': node,
+        'form': form,
+        'links': node.links.all(),
+        'linkform': linkform,
+        'country': node.country,
+        'parents_list': node.ancestors,
+        'supplements_list': node.supplements.all()
+        })
 
 def node_child(request, parent_id, nodetype_id):
     if request.method == 'POST':
@@ -155,16 +217,44 @@ def declaration_edit(request, declaration_id):
             saved = form.save()
             return redirect('node', node_id=dec.node.id)
     # Show form
-    return render(request, 'govtrack/declare.html', {'action': 'edit', 'declaration': dec, 'form': form, 'node': dec.node})
+    return render(request, 'govtrack/declare.html', {
+        'action': 'edit',
+        'declaration': dec,
+        'form': form,
+        'node': dec.node,
+        'links': dec.links.all(),
+        })
 
-# TODO integrate these into edit path
-# so we can't delete just by sending a POST
+# API methods
 def nodetype_del(request, nodetype_id):
-    nodetype = NodeType.objects.get(id=nodetype_id)
-    nodetype.delete()
-    return redirect('country', country_id=nodetype.country.id)
+    status=403
+    if request.user.is_authenticated:
+        status=200
+        nodetype = get_object_or_404(NodeType, pk=nodetype_id)
+        nodetype.delete()
+    return HttpResponse(status=status)
 
 def node_del(request, node_id):
-    node = Node.objects.get(id=node_id)
-    node.delete()
-    return redirect('country', country_id=node.country.id)
+    status=403
+    if request.user.is_authenticated:
+        status=200
+        node = get_object_or_404(Node, pk=node_id)
+        node.delete()
+    return HttpResponse(status=status)
+
+def declaration_del(request, declaration_id):
+    status=403
+    if request.user.is_authenticated:
+        status=200
+        declaration = get_object_or_404(Declaration, pk=declaration_id)
+        declaration.delete()
+    return HttpResponse(status=status)
+
+def link_del(request, link_id):
+    status=403
+    if request.user.is_authenticated:
+        status=200
+        link = get_object_or_404(Link, pk=link_id)
+        link.delete()
+    return HttpResponse(status=status)
+
